AWSBraket

API References

AWSBraket.DeviceMenu — Method

DeviceMenu([devices, regions]; pagesize::Int=5)

Create a DeviceMenu to display available devices interactively in the Julia REPL.

Arguments

devices::Vector{Schema.DeviceInfo}: optional, a list of devices.

regions::Vector{String}: optional, a list of regions corresponding to the device list.

AWSBraket.cancel_quantum_task — Method

cancel_quantum_task(client_token::String, task_arn::String)

Cancel quantum task given by client_token and its task_arn.

AWS Service Configuration

You can specify custom configuration via aws_config keyword argument, to select another regions or use a different account, default is AWS.global_aws_config(), e.g you can choose a different region in this config

using AWS
config = AWSConfig(;region="us-west-1")

See the AWS Julia Interface documentation for more advanced usage.

AWSBraket.create_quantum_task — Method

create_quantum_task(;kw...)

Create a quantum task in braket service.

Required Keyword Arguments

program::Schema.Program: the program one wants to execute.

device_arn::String: device arn.

bucket::String: S3 bucket to store the results in.

folder::String: S3 bucket folder.

Optional Keyword Arguments

disable_qubit_rewiring::Bool: disable qubit rewiring in braket service, default is false.

device_parameters: device parameters, such as Schema.IonqDeviceParameters, Schema.RigettiDeviceParameters, default is inferred from previous arguments.

nshots: number of shots, default is 100.

client_token: a UUID for the client token, will generate one by default.

tags::Dict{String, String}: a list of tags you would to attach to this task.

AWS Service Configuration

You can specify custom configuration via aws_config keyword argument, to select another regions or use a different account, default is AWS.global_aws_config(), e.g you can choose a different region in this config

using AWS
config = AWSConfig(;region="us-west-1")

See the AWS Julia Interface documentation for more advanced usage.

AWSBraket.get_device — Method

get_device(arn::String; aws_config=AWS.global_aws_config())

Retrieves the devices available in Amazon Braket.

Arguments

arn: The ARN of the device to retrieve.

AWS Service Configuration

You can specify custom configuration via aws_config keyword argument, to select another regions or use a different account, default is AWS.global_aws_config(), e.g you can choose a different region in this config

using AWS
config = AWSConfig(;region="us-west-1")

See the AWS Julia Interface documentation for more advanced usage.

AWSBraket.get_quantum_task — Method

get_quantum_task(task_arn::String)

Get the quantum task from task_arn.

AWS Service Configuration

You can specify custom configuration via aws_config keyword argument, to select another regions or use a different account, default is AWS.global_aws_config(), e.g you can choose a different region in this config

using AWS
config = AWSConfig(;region="us-west-1")

See the AWS Julia Interface documentation for more advanced usage.

AWSBraket.make_device_parameters — Method

make_device_parameters(program::Schema.Program, arn::String, disable_qubit_rewiring::Bool)

Create device parameters from given program, device arn and disable_qubit_rewiring option.

AWSBraket.search_devices — Function

search_devices(filters=[]; max_results::Maybe{Int}=nothing, next_token::Maybe{String}=nothing, aws_config=AWS.global_aws_config())

Searches for devices using the specified filters.

Arguments

filters: The filter values to use to search for a device.

Keyword Arguments (Optional)

max_results: The maximum number of results to return in the response.

next_token: A token used for pagination of results returned in the response. Use the token returned from the previous request continue results where the previous request ended.

AWS Service Configuration

You can specify custom configuration via aws_config keyword argument, to select another regions or use a different account, default is AWS.global_aws_config(), e.g you can choose a different region in this config

using AWS
config = AWSConfig(;region="us-west-1")

See the AWS Julia Interface documentation for more advanced usage.